refactor : main controller에서 출력에 관한 책임을 하위 output controller에서 위임
